Fix logging in going to the twitch landing page
Closes #78

For whatever reason, if Twitch thinks we're a mobile user it'll redirect us to a mobile OAuth page (m.twitch.tv/login?client_id=...) and then that redirects us again to the standard login (www.twitch.tv/login) which is why we go to the landing page instead of our redirect uri. By changing the mobile UA into a desktop UA, Twitch doesn't think we're on mobile and doesn't redirect us to the mobile OAuth page.